Create gameplay class and refactor runner
